+++ api/controller/project.controller.ts
@@ -6,6 +6,9 @@ import AppError from "../utils/app-error";
 import { StatusCodes } from "http-status-codes";
 import Project from "../model/project.model";
 import slugify from "slugify";
+import { createNotificationService } from "../service/notification.service";
+import { generateNotificationName } from "../constants/notification.constant";
+import { NotificationTypeEnum } from "../model/notification.model";
 
 const getProjects = catchAsync(
   async (req: Request, res: Response, next: NextFunction) => {
@@ -13,7 +16,7 @@ const getProjects = catchAsync(
       Project.find({
         user: req.userId,
       }),
-      req.query
+      req.query,
     )
       .filter()
       .sort()
@@ -32,7 +35,7 @@ const getProjects = catchAsync(
         projects,
       },
     });
-  }
+  },
 );
 
 const getProject = catchAsync(
@@ -44,7 +47,7 @@ const getProject = catchAsync(
 
     if (!project) {
       return next(
-        new AppError(`No project found with the id ${req.params.id}`, 404)
+        new AppError(`No project found with the id ${req.params.id}`, 404),
       );
     }
 
@@ -54,7 +57,7 @@ const getProject = catchAsync(
         project,
       },
     });
-  }
+  },
 );
 
 const createProject = catchAsync(
@@ -68,8 +71,8 @@ const createProject = catchAsync(
       return next(
         new AppError(
           `Project with the name '${req.body.name}' already exists.`,
-          StatusCodes.CONFLICT
-        )
+          StatusCodes.CONFLICT,
+        ),
       );
 
     const project = await Project.create({
@@ -86,7 +89,7 @@ const createProject = catchAsync(
         project,
       },
     });
-  }
+  },
 );
 
 const deleteProject = catchAsync(
@@ -103,16 +106,27 @@ const deleteProject = catchAsync(
 
     if (!project) {
       return next(
-        new AppError(`No project found with the id ${req.params.id}`, 404)
+        new AppError(`No project found with the id ${req.params.id}`, 404),
       );
     }
+
+    await createNotificationService({
+      name: generateNotificationName(
+        NotificationTypeEnum.PROJECT_DELETED,
+        project.name,
+      )!,
+      data: project,
+      value: project.id,
+      type: NotificationTypeEnum.PROJECT_DELETED,
+      user: req.userId!,
+    });
 
     res.status(StatusCodes.NO_CONTENT).json({
       status: "success",
       message: "Project successfully deleted.",
       data: null,
     });
-  }
+  },
 );
 
 const updateProject = catchAsync(
@@ -125,10 +139,21 @@ const updateProject = catchAsync(
       return next(
         new AppError(
           `No project found with the id ${req.params.id}`,
-          StatusCodes.NOT_FOUND
-        )
+          StatusCodes.NOT_FOUND,
+        ),
       );
     }
+
+    await createNotificationService({
+      name: generateNotificationName(
+        NotificationTypeEnum.PROJECT_UPDATED,
+        project.name,
+      )!,
+      data: project,
+      value: project.id,
+      type: NotificationTypeEnum.PROJECT_UPDATED,
+      user: req.userId!,
+    });
 
     // const project = await Project.findOneAndUpdate(
     //   {
@@ -154,7 +179,7 @@ const updateProject = catchAsync(
       message: "Project successfully updated.",
       data: { project },
     });
-  }
+  },
 );
 
 const addProjectToFavorites = catchAsync(
@@ -165,7 +190,7 @@ const addProjectToFavorites = catchAsync(
     });
     if (!existingProject) {
       return next(
-        new AppError(`No project found with the id ${req.params.id}`, 404)
+        new AppError(`No project found with the id ${req.params.id}`, 404),
       );
     }
 
@@ -176,7 +201,7 @@ const addProjectToFavorites = catchAsync(
       },
       {
         favorite: true,
-      }
+      },
     );
 
     res.status(StatusCodes.OK).json({
@@ -184,7 +209,7 @@ const addProjectToFavorites = catchAsync(
       message: "Project successfully added to favorites.",
       data: { project },
     });
-  }
+  },
 );
 
 const removeProjectFromFavorites = catchAsync(
@@ -195,7 +220,7 @@ const removeProjectFromFavorites = catchAsync(
     });
     if (!existingProject) {
       return next(
-        new AppError(`No project found with the id ${req.params.id}`, 404)
+        new AppError(`No project found with the id ${req.params.id}`, 404),
       );
     }
 
@@ -206,14 +231,14 @@ const removeProjectFromFavorites = catchAsync(
       },
       {
         favorite: false,
-      }
+      },
     );
 
     res.status(204).json({
       status: "success",
       message: "Project successfully removed from favorites",
     });
-  }
+  },
 );
 
 export {