+++ api/index.ts
@@ -14,7 +14,7 @@ process.on("uncaughtException", (error: Error) => {
 const server = app.listen(port, async () => {
   console.log(`[${node_env}] Server is running on port ${port}`);
   await connectToDatabase();
-  initializeScheduler();
+  await initializeScheduler();
 });
 
 process.on("unhandledRejection", (error: Error) => {